pyproject.toml
[tool.poetry]
name = "drfproject"
version = "0.1.0"
description = ""
authors = ["Andrey Tumerkin <tumerrkin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
ipython = "^8.30.0"
psycopg2-binary = "^2.9.10"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
